HHH-13051 - Refactored strategy solution.
