+++ hibernate-envers/src/main/java/org/hibernate/envers/strategy/AuditStrategy.java
@@ -9,26 +9,21 @@ package org.hibernate.envers.strategy;
 import java.io.Serializable;
 
 import org.hibernate.Session;
-import org.hibernate.service.ServiceRegistry;
 import org.hibernate.envers.boot.internal.EnversService;
 import org.hibernate.envers.configuration.internal.AuditEntitiesConfiguration;
-import org.hibernate.envers.configuration.internal.GlobalConfiguration;
-import org.hibernate.envers.internal.entities.PropertyData;
 import org.hibernate.envers.internal.entities.mapper.PersistentCollectionChangeData;
-import org.hibernate.envers.internal.entities.mapper.relation.MiddleComponentData;
-import org.hibernate.envers.internal.entities.mapper.relation.MiddleIdData;
-import org.hibernate.envers.internal.tools.query.Parameters;
-import org.hibernate.envers.internal.tools.query.QueryBuilder;
-import org.dom4j.Element;
 
 /**
  * Behaviours of different audit strategy for populating audit data.
+ *
+ * @deprecated (since 5.4), use {@link org.hibernate.envers.strategy.spi.AuditStrategy} instead.
  *
  * @author Stephanie Pau
  * @author Adam Warski (adam at warski dot org)
  * @author Chris Cranford
  */
-public interface AuditStrategy {
+@Deprecated
+public interface AuditStrategy extends org.hibernate.envers.strategy.spi.AuditStrategy {
 	/**
 	 * Perform the persistence of audited data for regular entities.
 	 *
@@ -58,23 +53,6 @@ public interface AuditStrategy {
 		);
 	}
 
-	/**
-	 * Perform the persistence of audited data for regular entities.
-	 *
-	 * @param session Session, which can be used to persist the data.
-	 * @param entityName Name of the entity, in which the audited change happens
-	 * @param auditEntitiesConfiguration The audit entity configuration.
-	 * @param id Id of the entity.
-	 * @param data Audit data to persist.
-	 * @param revision Current revision data.
-	 */
-	void perform(
-			Session session,
-			String entityName,
-			AuditEntitiesConfiguration auditEntitiesConfiguration,
-			Serializable id,
-			Object data,
-			Object revision);
 
 	/**
 	 * Perform the persistence of audited data for collection ("middle") entities.
@@ -103,131 +81,5 @@ public interface AuditStrategy {
 				persistentCollectionChangeData,
 				revision
 		);
-	}
-
-	/**
-	 * Perform the persistence of audited data for collection ("middle") entities.
-	 *
-	 * @param session Session, which can be used to persist the data.
-	 * @param entityName Name of the entity, in which the audited change happens.
-	 * @param propertyName The name of the property holding the persistent collection
-	 * @param auditEntitiesConfiguration audit entity configuration
-	 * @param persistentCollectionChangeData Collection change data to be persisted.
-	 * @param revision Current revision data
-	 */
-	void performCollectionChange(
-			Session session,
-			String entityName,
-			String propertyName,
-			AuditEntitiesConfiguration auditEntitiesConfiguration,
-			PersistentCollectionChangeData persistentCollectionChangeData,
-			Object revision);
-
-	/**
-	 * Update the rootQueryBuilder with an extra WHERE clause to restrict the revision for a two-entity relation.
-	 * This WHERE clause depends on the AuditStrategy, as follows:
-	 * <ul>
-	 * <li>For {@link DefaultAuditStrategy} a subquery is created:
-	 * <p><code>e.revision = (SELECT max(...) ...)</code></p>
-	 * </li>
-	 * <li>for {@link ValidityAuditStrategy} the revision-end column is used:
-	 * <p><code>e.revision <= :revision and (e.endRevision > :revision or e.endRevision is null)</code></p>
-	 * </li>
-	 * </ul>
-	 *
-	 * @param globalCfg the {@link GlobalConfiguration}
-	 * @param rootQueryBuilder the {@link QueryBuilder} that will be updated
-	 * @param parameters root parameters to which restrictions shall be added
-	 * @param revisionProperty property of the revision column
-	 * @param revisionEndProperty property of the revisionEnd column (only used for {@link ValidityAuditStrategy})
-	 * @param addAlias {@code boolean} indicator if a left alias is needed
-	 * @param idData id-information for the two-entity relation (only used for {@link DefaultAuditStrategy})
-	 * @param revisionPropertyPath path of the revision property (only used for {@link ValidityAuditStrategy})
-	 * @param originalIdPropertyName name of the id property (only used for {@link ValidityAuditStrategy})
-	 * @param alias1 an alias used for subquery (only used for {@link ValidityAuditStrategy})
-	 * @param alias2 an alias used for subquery (only used for {@link ValidityAuditStrategy})
-	 * @param inclusive indicates whether revision number shall be treated as inclusive or exclusive
-	 */
-	void addEntityAtRevisionRestriction(
-			GlobalConfiguration globalCfg,
-			QueryBuilder rootQueryBuilder,
-			Parameters parameters,
-			String revisionProperty,
-			String revisionEndProperty,
-			boolean addAlias,
-			MiddleIdData idData,
-			String revisionPropertyPath,
-			String originalIdPropertyName,
-			String alias1,
-			String alias2,
-			boolean inclusive);
-
-	/**
-	 * Update the rootQueryBuilder with an extra WHERE clause to restrict the revision for a middle-entity
-	 * association. This WHERE clause depends on the AuditStrategy, as follows:
-	 * <ul>
-	 * <li>For {@link DefaultAuditStrategy} a subquery is created:
-	 * <p><code>e.revision = (SELECT max(...) ...)</code></p>
-	 * </li>
-	 * <li>for {@link ValidityAuditStrategy} the revision-end column is used:
-	 * <p><code>e.revision <= :revision and (e.endRevision > :revision or e.endRevision is null)</code></p>
-	 * </li>
-	 * </ul>
-	 *
-	 * @param rootQueryBuilder the {@link QueryBuilder} that will be updated
-	 * @param parameters root parameters to which restrictions shall be added
-	 * @param revisionProperty property of the revision column
-	 * @param revisionEndProperty property of the revisionEnd column (only used for {@link ValidityAuditStrategy})
-	 * @param addAlias {@code boolean} indicator if a left alias is needed
-	 * @param referencingIdData id-information for the middle-entity association (only used for {@link DefaultAuditStrategy})
-	 * @param versionsMiddleEntityName name of the middle-entity
-	 * @param eeOriginalIdPropertyPath name of the id property (only used for {@link ValidityAuditStrategy})
-	 * @param revisionPropertyPath path of the revision property (only used for {@link ValidityAuditStrategy})
-	 * @param originalIdPropertyName name of the id property (only used for {@link ValidityAuditStrategy})
-	 * @param alias1 an alias used for subqueries (only used for {@link DefaultAuditStrategy})
-	 * @param inclusive indicates whether revision number shall be treated as inclusive or exclusive
-	 * @param componentDatas information about the middle-entity relation
-	 */
-	void addAssociationAtRevisionRestriction(
-			QueryBuilder rootQueryBuilder,
-			Parameters parameters,
-			String revisionProperty,
-			String revisionEndProperty,
-			boolean addAlias,
-			MiddleIdData referencingIdData,
-			String versionsMiddleEntityName,
-			String eeOriginalIdPropertyPath,
-			String revisionPropertyPath,
-			String originalIdPropertyName,
-			String alias1,
-			boolean inclusive,
-			MiddleComponentData... componentDatas);
-
-	/**
-	 * Method to add additional Columns to the aud-tables.
-	 *
-	 * @param anyMapping the class mapping
-	 * @param revMapping the mapping of the revInfo-relation
-	 * @param auditEntitiesConfiguration the configuration of the audit entities
-	 */
-	default void addAdditionalColumns(
-			Element anyMapping,
-			Element revMapping,
-			AuditEntitiesConfiguration auditEntitiesConfiguration) {
-		//For backward compatibility
-	}
-
-	/**
-	 * Method do add additional initialization behavior
-	 *
-	 * @param revisionInfoClass the class of the revision info
-	 * @param revisionInfoTimestampData the timestamp of the revision info
-	 * @param serviceRegistry the service registry
-	 */
-	default void initialize(
-			Class<?> revisionInfoClass,
-			PropertyData revisionInfoTimestampData,
-			ServiceRegistry serviceRegistry) {
-		//For backward compatibility
 	}
 }
